backend/celery-worker/requirements.txt
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
amqp==5.2.0
